+++ nflow-engine/src/main/java/io/nflow/engine/internal/executor/WorkflowStateProcessor.java
@@ -292,6 +292,9 @@ class WorkflowStateProcessor implements Runnable {
         .setState(execution.getNextState()) //
         .setRetries(execution.isRetry() ? execution.getRetries() + 1 : 0);
     int saveRetryCount = 0;
+    if (execution.getNewBusinessKey() != null) {
+      instanceBuilder.setBusinessKey(execution.getNewBusinessKey());
+    }
     while (true) {
       try {
         return persistWorkflowInstanceState(execution, instance.stateVariables, actionBuilder, instanceBuilder);